new portal component for upset
