+++ apps/synapse-portal-framework/src/components/index.ts
@@ -17,6 +17,7 @@ import ELBrowseToolsPage from './eliteportal/ELBrowseToolsPage'
 import ELContributeYourData from './eliteportal/ELContributeYourData'
 import ELGettingStarted from './eliteportal/ELGettingStarted'
 import AllALSSlat from './ampals/ALLALSSlat'
+import AMPALSExploretheData from './ampals/AMPALSExploreTheData'
 import ELSupportedByNIH from './eliteportal/ELSupportedByNIH'
 import ARKWelcomePage from './arkportal/ARKWelcomePage'
 
@@ -55,6 +56,7 @@ const PortalComponents = {
   ELContributeYourData,
   ELGettingStarted,
   AllALSSlat,
+  AMPALSExploretheData,
   ELSupportedByNIH,
   ARKWelcomePage,
   GenieHomePageHeader,
